More realistic benchmark
But less 'isolated' since exercising both client and server
