+++ akka-http-bench-jmh/src/main/scala/akka/http/impl/engine/http2/H2ServerProcessingBenchmark.scala
@@ -4,54 +4,55 @@
 
 package akka.http.impl.engine.http2
 
-import java.util.concurrent.{ CountDownLatch, TimeUnit }
-import scala.collection.immutable
-import scala.concurrent.Await
-import scala.concurrent.duration._
 import akka.actor.ActorSystem
 import akka.http.CommonBenchmark
-import akka.http.scaladsl.client.RequestBuilding.Get
-import akka.http.scaladsl.model.HttpEntity.LastChunk
-import akka.http.scaladsl.model.headers.RawHeader
-import akka.http.scaladsl.model.{ ContentTypes, HttpEntity, HttpRequest, HttpResponse }
+import akka.http.impl.engine.http2.FrameEvent.HeadersFrame
+import akka.http.impl.engine.http2.framing.FrameRenderer
+import akka.http.scaladsl.model.HttpResponse
+import akka.http.scaladsl.settings.ServerSettings
 import akka.stream.ActorMaterializer
+import akka.stream.TLSProtocol.{ SslTlsInbound, SslTlsOutbound }
 import akka.stream.scaladsl.{ Flow, Keep, Sink, Source }
+import akka.util.ByteString
 import com.typesafe.config.ConfigFactory
 import org.openjdk.jmh.annotations._
 
-import scala.util.{ Failure, Success }
+import java.util.concurrent.{ CountDownLatch, TimeUnit }
+import scala.concurrent.{ Await, Future }
+import scala.concurrent.duration._
 
-class H2ServerProcessingBenchmark extends CommonBenchmark {
+class H2ServerSwitchBenchmark extends CommonBenchmark {
+  // Obtained by converting the input request bytes from curl with --http2-prior-knowledge
+  def request(streamId: Int) =
+    FrameRenderer.render(HeadersFrame(streamId, endStream = true, endHeaders = true, HPackSpecExamples.C41FirstRequestWithHuffman, None))
 
-  @Param(Array("closedelimited", "chunked"))
-  var responsetype: String = _
+  var response: HttpResponse = HPackSpecExamples.FirstResponse
 
-  var response: HttpResponse = _
-
-  var httpFlow: Flow[HttpRequest, HttpResponse, Any] = _
+  var httpFlow: Flow[ByteString, ByteString, Any] = _
   implicit var system: ActorSystem = _
   implicit var mat: ActorMaterializer = _
+
+  val packedResponse = ByteString(-62, -63, -64, -65, -66)
 
   val numRequests = 10000
 
   @Benchmark
   @OperationsPerInvocation(10000) // should be same as numRequest
   def benchRequestProcessing(): Unit = {
-    implicit val ec = system.dispatcher
-
     val latch = new CountDownLatch(numRequests)
 
     val requests =
-      Source.repeat(Get("/")).take(numRequests)
+      Source(Http2Protocol.ClientConnectionPreface +: Range(0, numRequests).map(i => request(1 + 2 * i)))
         .concatMat(Source.maybe)(Keep.right)
 
     val (in, done) =
       requests
         .viaMat(httpFlow)(Keep.left)
         .toMat(Sink.foreach(res => {
-          res.entity.dataBytes.runWith(Sink.ignore).onComplete {
-            case Success(_) => latch.countDown()
-            case Failure(e) => throw e
+          // Skip headers/settings frames etc
+          if (res.containsSlice(HPackSpecExamples.C61FirstResponseWithHuffman)
+            || res.containsSlice(packedResponse)) {
+            latch.countDown()
           }
         }))(Keep.both)
         .run()
@@ -64,12 +65,6 @@ class H2ServerProcessingBenchmark extends CommonBenchmark {
 
   @Setup
   def setup(): Unit = {
-    response = responsetype match {
-      case "closedelimited" => HPackSpecExamples.FirstResponse
-      case "chunked" => HPackSpecExamples.FirstResponse.withEntity(
-        HttpEntity.Chunked(ContentTypes.NoContentType, Source.single(LastChunk(trailer = immutable.Seq(RawHeader("grpc-status", "9")))))
-      )
-    }
     val config =
       ConfigFactory.parseString(
         s"""
@@ -81,10 +76,18 @@ class H2ServerProcessingBenchmark extends CommonBenchmark {
         .withFallback(ConfigFactory.load())
     system = ActorSystem("AkkaHttpBenchmarkSystem", config)
     mat = ActorMaterializer()
+    val settings = implicitly[ServerSettings]
+    val log = system.log
     implicit val ec = system.dispatcher
-    httpFlow = Http2Blueprint.handleWithStreamIdHeader(1)(req => {
-      req.discardEntityBytes().future.map(_ => response)
-    })(system.dispatcher)
+    val http1 = Flow[SslTlsInbound].mapAsync(1)(_ => {
+      Future.failed[SslTlsOutbound](new IllegalStateException("Failed h2 detection"))
+    })
+    val http2 =
+      Http2Blueprint.handleWithStreamIdHeader(1)(req => {
+        req.discardEntityBytes().future.map(_ => response)
+      })(system.dispatcher)
+        .join(Http2Blueprint.serverStackTls(settings, log, NoOpTelemetry))
+    httpFlow = Http2.priorKnowledge(http1, http2)
   }
 
   @TearDown
